getAllLocations API added

+++ src/main/java/edu/miu/cs/cs544/a202001/wind/controller/LocationController.java
@@ -18,7 +18,7 @@ public class LocationController {
     }
 
     @GetMapping(value = "/{id}")
-    public Location getLocation(@PathVariable int id) {
+    public Location getLocation(@PathVariable Long id) {
         return locationService.getLocationById(id);
     }
