+++ src/main/java/edu/miu/cs/cs544/a202001/wind/repository/ILocationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ILocationRepository extends JpaRepository<Location, Integer> {
+public interface ILocationRepository extends JpaRepository<Location, Long> {
 }
